Evaluation site files
This website is made for the judges and evaluators. The main purpose of this kind of site is to eliminate the chances of biasness while evaluating the solutions. The judges/evaluators do not get the know the participant name or any other detail regarding the participant. They evaluate and provide the marks in the fields provided. Rest of the work is done by the module itself.